feat: add rectangle
